+++ src/analytics.js
@@ -16,7 +16,7 @@ module.exports = Analytics;
 
 
 function Analytics (Providers) {
-  this.VERSION = '0.7.0';
+  this.VERSION = '0.7.1';
 
   var self = this;
   // Loop through and add each of our `Providers`, so they can be initialized