+++ test/providers/intercom.js
@@ -1,6 +1,18 @@
 
 describe('Intercom', function () {
 
+  var options = {
+    appId     : 'abc123',
+    activator : '#someElement',
+  };
+
+  var userId = test.userId
+  ,   traits = clone(test.traits);
+
+  extend(traits, {
+    company : { id : '123' }
+  });
+
   describe('initialize', function () {
 
     it('should call ready', function () {
@@ -15,6 +27,14 @@ describe('Intercom', function () {
       expect(analytics.providers[0].options.appId).to.equal(test['Intercom']);
     });
 
+    it('should store expanded options', function () {
+
+      analytics.initialize({ 'Intercom' : options });
+
+      options.counter = true;
+
+      expect(analytics.providers[0].options).to.eql(options);
+    });
   });
 
 
@@ -27,8 +47,22 @@ describe('Intercom', function () {
       expect(window.intercomSettings).to.be(undefined);
 
       expect(window.Intercom).to.be(undefined);
-      analytics.identify(test.userId, test.traits);
+      analytics.identify(test.userId, traits);
       expect(window.intercomSettings).not.to.be(undefined);
+      expect(window.intercomSettings).to.eql({
+        created_at  : Math.floor(traits.created/1000),
+        app_id      : options.appId,
+        user_id     : userId,
+        company     : traits.company,
+        custom_data : traits,
+        user_hash   : undefined,
+        email       : traits.email,
+        name        : traits.name,
+        widget      : {
+          activator   : options.activator,
+          use_counter : options.counter
+        }
+      });
 
 
       // Once the Intercom library comes back, the array should be transformed.